Made it better to put in a map, made test to compare it
island.py: Made a staticmethod, function make_map_ready, for the make_a_map function. This function basically checks if the string is all good, if not it will raise a ValueError.
 I did some minor changes too for pep 8

test_island.py: 
made a new test, test_edges_water(). It tests if you get an error when you set  a letter other then 'W' along the edges